+++ plugin/src/main/kotlin/com/linecorp/intellij/plugins/armeria/explorer/ArmeriaRouteExplorerPanel.kt
@@ -90,7 +90,13 @@ class ArmeriaRouteExplorerPanel(
                         }
                     },
                 )
-                add(ArmeriaGenerateHttpRequestAction { selectedRouteFromTree() })
+                add(
+                    ArmeriaGenerateHttpRequestAction(
+                        selectedRouteProvider = { selectedRouteFromTree() },
+                        routesProvider = { filterRoutes(allRoutes()) },
+                        lastSyncedDocsBaseUrlProvider = { routeState.lastDocServiceBaseUrl },
+                    ),
+                )
                 add(ArmeriaExportOpenApiAction { filterRoutes(allRoutes()) })
                 add(ArmeriaGenerateTestMethodAction { selectedRouteFromTree() })
                 add(